refactor: 도메인 모델 - 도메인별 패키징
