+++ src/domain/models/index.ts
@@ -1,10 +1,6 @@
-export * from './cart-item.model';
-export * from './cart.model';
-export * from './order-item.model';
-export * from './order.model';
+export * from './cart';
+export * from './order';
 export * from './point.model';
-export * from './popular-product.model';
-export * from './product-stock.model';
-export * from './product.model';
+export * from './product';
 export * from './user.model';
 export * from './wallet.model';
