+++ src/domain/models/order/index.ts
@@ -1,0 +1,2 @@
+export * from './order-item.model';
+export * from './order.model';
