+++ src/domain/models/product/index.ts
@@ -1,0 +1,3 @@
+export * from './popular-product.model';
+export * from './product-stock.model';
+export * from './product.model';
