+++ src/domain/models/product/popular-product.model.ts
