+++ src/domain/models/product/product.model.ts
